Create publisher snaps blueprint
