+++ webapp/publisher/snaps/logic.py
